+++ styles.css
@@ -303,7 +303,8 @@
     border: none;
     background: none;
     cursor: pointer;
-    font-size: 16px;
+    font-size: 13px;
+    text-transform: uppercase;
     color: var(--text-muted);
     transition: color 0.3s ease;
 }
